#15 fix: participant 수 구할 때 active인 것만 카운팅하도록 수정

+++ src/main/java/com/hatcher/haemo/notification/dto/NotificationDto.java
@@ -4,6 +4,6 @@ public record NotificationDto(Long notificationIdx,
                               String name,
                               String contactUrl,
                               String leader,
-                              long participantNumber,
+                              int participantNumber,
                               Integer participantLimit,
                               String description) {}
